+++ generate/build-declarations.js
@@ -141,6 +141,7 @@ ${brandedElements.map(el => `export class ${el.pascalTag}Tag extends ${el.return
  */
 export interface ReadonlySignal<T> {
   get(): T;
+  readonly value: T;
   stop(): void;
   transform<U>(fn: (value: T) => U): ReadonlySignal<U>;
 }
@@ -153,6 +154,7 @@ export interface ReadonlySignal<T> {
 export class Signal<T> implements ReadonlySignal<T> {
   private constructor();
   get(): T;
+  readonly value: T;
   set(valueOrFn: T | ((current: T) => T)): void;
   stop(): void;
   transform<U>(fn: (value: T) => U): ReadonlySignal<U>;