+++ types.d.ts
@@ -83,6 +83,7 @@ export class UlTag extends ContentTag { private readonly _k: 'ul' }
  */
 export interface ReadonlySignal<T> {
   get(): T;
+  readonly value: T;
   stop(): void;
   transform<U>(fn: (value: T) => U): ReadonlySignal<U>;
 }
@@ -95,6 +96,7 @@ export interface ReadonlySignal<T> {
 export class Signal<T> implements ReadonlySignal<T> {
   private constructor();
   get(): T;
+  readonly value: T;
   set(valueOrFn: T | ((current: T) => T)): void;
   stop(): void;
   transform<U>(fn: (value: T) => U): ReadonlySignal<U>;